third_party/WebKit/LayoutTests/imported/csswg-test/css-writing-modes-3/writing-mode-parsing-svg1-001-expected.txt

This is a testharness.js-based test.
FAIL -webkit-writing-mode: lr assert_equals: expected "lr" but got "lr-tb"
FAIL -webkit-writing-mode: rl assert_equals: expected "rl" but got "lr-tb"
FAIL -webkit-writing-mode: tb assert_equals: expected "tb" but got "lr-tb"
Harness: the test ran to completion.
